clean interfaces
